it now compiles on OS X, but does it work?
git-svn-id: svn://svn.icculus.org/netradiant/trunk@169 61c419a2-8eb2-4b30-bcec-8cead039b335

+++ radiant/texwindow.cpp
@@ -1917,7 +1917,7 @@ void TextureBrowser_searchTags()
     {
 	   size_t shaders_found = g_TextureBrowser.m_found_shaders.size();
 
-       globalOutputStream() << "Found " << shaders_found << " textures and shaders with " << tags_searched << "\n";
+       globalOutputStream() << "Found " << (unsigned int)shaders_found << " textures and shaders with " << (unsigned int)tags_searched << "\n";
 	   ScopeDisableScreenUpdates disableScreenUpdates("Searching...", "Loading Textures");
 
 	  std::set<CopiedString>::iterator iter;
